WIP: draw text

+++ src/image.h
@@ -13,19 +13,18 @@ typedef enum ImageChannel {
     IMAGE_CHANNEL_A,
 } ImageChannel;
 
-// The image is processed with pre-multiplied alpha.
 typedef struct Image {
     ImageSource source;
     ImageChannel channel;
     const char *name;
     int width;
     int height;
-    size_t stride;
+    int stride;
     unsigned char *data;
 } Image;
 
 extern Image *LoadImageFromFilename(const char *filename);
-extern Image *LoadImageFromGrayBitmap(int width, int height, size_t stride, const unsigned char *data);
+extern Image *LoadImageFromGrayBitmap(int width, int height, int stride, const unsigned char *data);
 extern void DestroyImage(Image **image);
 
 #endif // RTD_IMAGE_H
